day 18 solves

+++ Challenge2023/Day18/Models/Node.cs
@@ -1,15 +1,15 @@
 ﻿namespace Challenge2023.Day18.Models
 {
-    internal class Node(string direction, int steps, string color, int x, int y)
+    internal class Node(string direction, long steps, string color, long x, long y)
     {
         public string Direction { get; } = direction;
         
-        public int Steps { get; } = steps;
+        public long Steps { get; } = steps;
         
         public string ColorHex { get; } = color;
 
-        public int X { get; } = x;
+        public long X { get; } = x;
 
-        public int Y { get; } = y;
+        public long Y { get; } = y;
     }
 }
